requirements.txt
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
boto3==1.42.14
botocore==1.42.14
brotli==1.2.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.2.4
django-anymail==13.1
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-storages==1.14.6
django-unfold==0.49.1
django_paystack==1.0.0
djangorestframework==3.15.1
filelock==3.15.4
fonttools==4.61.0
freetype-py==2.5.1
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jmespath==1.0.1
lxml==6.0.2
msgpack==1.1.2
oscrypto==1.3.0
packaging==24.0
paystack==1.5.0
paystackapi==2.1.3
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.22
pydyf==0.12.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pyOpenSSL==25.3.0
pypaystack2==2.0.3
pypdf==6.4.1
PyPDF2==3.0.1
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
reportlab==4.4.6
requests==2.32.3
rlPyCairo==0.4.0
s3transfer==0.16.0
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
svglib==1.6.0
sweetify==2.3.1
tinycss2==1.5.1
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.2.2
virtualenv==20.26.3
weasyprint==67.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.17
zope.interface==8.0.1
zopfli==0.4.0
